Clarify input dimensions for pcolormesh
The tutorial states that the `x`, `y`, and `data` inputs to pcolormesh should be of the same size, but that's not correct.  As per the `Basemap.pcolormesh` docstring, `x` and `y` should have dimensions one greater than those of data; the code will still run if they have the same size but the last row and column get dropped.